Fixes #18 Azure does not unescape site extension filenames Also improves speed and reliability of install script

+++ 2714x86/content/install.cmd
@@ -1,3 +1,18 @@
+set SRC=Python27
+
+if not exist %SRC% (
+    if not exist D:\home\%SRC% (
+        echo Cannot reinstall previously installed site extension.
+        echo Please remove %~dp0 and reinstall.
+        exit /B 1
+    )
+    exit /B 0
+)
 taskkill /F /IM python.exe
-if exist D:\home\Python27 rmdir /Q /S D:\home\Python27
-xcopy /s Python27 D:\home\Python27\
+if exist D:\home\%SRC% rmdir /Q /S D:\home\%SRC%
+move %SRC% D:\home\%SRC%
+
+if exist D:\home\%SRC% (
+    ren D:\home\%SRC%\Lib\site-packages\setuptools\script%%%%20(dev).tmpl "script (dev).tmpl"
+    ren D:\home\%SRC%\Lib\site-packages\setuptools\launcher%%%%20manifest.xml "launcher manifest.xml"
+)
